buy: provide a first integration of Pocket widget
This commit is a first step in the integration process of Pocket widget
into the buy workflow.

The widget is embedded in a dedicated iframe in `pocket.tsx`, while
a drafted `exchange.tsx` allows the user to choose which exchange
(moonpay or pocket) should be used in the purchase process. The iframe
integration includes an address verification step similar to the one
used with AOPP flow. In order to do this, a typescript `request-address`
library has been added to the `package.json`.

Backend side, a new `pocket.go` contains the functions needed to verify
coin support and iframe URL, while the code dedicated to the address
verification is in `backend.go`.